pyproject.toml
[tool.poetry]
name = "rattletat.com"
version = "0.1.0"
description = ""
authors = ["Michael Brauweiler <michael.brauweiler@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.8"
django-environ = "^0.4.5"
psycopg2 = "^2.8.5"
django-model-utils = "^4.0.0"
argon2_cffi = "^20.1.0"
django-markdownx = "^3.0.1"
bleach = "^3.1.5"
markdown-prism = "^1.0.1"
django-next-prev = "^1.0.1"

[tool.poetry.dev-dependencies]
ipython = "^7.15.0"
flake8 = "^3.8.2"
black = "^19.10b0"
django-debug-toolbar = "^2.2"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
